+++ go/clustermetadata/topo/cell_location.go
