+++ go/clustermetadata/topo/conn.go
