+++ go/clustermetadata/topo/errors.go
